backend/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Dependency Map API (FastAPI)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic-settings>=2.6",
  "pyjwt[crypto]>=2.8",
  "httpx>=0.28",
  "supabase>=2.10",
  "networkx>=3.2",
  "celery[redis]>=5.4",
  "redis>=5.0",
  "slowapi>=0.1.9",
  "tree-sitter>=0.21",
  "tree-sitter-typescript>=0.23",
  "tree-sitter-javascript>=0.23",
  "tree-sitter-python>=0.23",
  "tree-sitter-java>=0.23",
  "openai>=1.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.8",
  "mypy>=1.13",
  "pytest>=8.0",
]
ml = [
  "torch>=2.2",
  "torch-geometric>=2.5",
  "sentence-transformers>=3.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "cpg_builder"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
